Make getNoiseShape to be a utility function for dropout op
