+++ src/tests.ts
@@ -56,6 +56,7 @@ import './ops/conv2d_transpose_test';
 import './ops/conv3d_test';
 import './ops/conv_util_test';
 import './ops/dropout_test';
+import './ops/dropout_util_test';
 import './ops/fused_test';
 import './ops/gather_nd_test';
 import './ops/image_ops_test';
